% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writebatchfile.R
\name{writebatchfile}
\alias{writebatchfile}
\title{A Function to Write Batch for Running FVS Keyfiles}
\usage{
writebatchfile(keyfile, variant, path_to_bat)
}
\arguments{
\item{keyfile}{name of the .key file.}

\item{variant}{Specify the FVS variant. Character string, two letters,
lower case. See example.}

\item{path_to_bat}{Character string with .bat extension, file path where
.bat file will be created.

#' @export}
}
\description{
This function creates a batch file to run FVS. Keyfiles can be created using
writekeyfile.
}
\examples{
\dontrun{
writebatchfile(example.key, "nc", "temp/path/file.bat")
}
}
